requirements.txt
fastapi
uvicorn
rembg
Pillow

